added CohortStagingToDataCatalogue strategy, removed and renamed multiple files
